requirements.txt
requests
aiohttp
tenacity
beautifulsoup4
lxml
playwright
flask
flask-compress
flask-cors
gunicorn
python-dotenv
psycopg2-binary
bcrypt
PyJWT
tzdata
